emit

+++ plugins/RmFilePlugin.js
@@ -1,3 +1,11 @@
+/*
+ * @Author: your name
+ * @Date: 2022-02-09 19:21:53
+ * @LastEditTime: 2022-02-09 21:49:58
+ * @LastEditors: your name
+ * @Description: 打开koroFileHeader查看配置 进行设置: https://github.com/OBKoro1/koro1FileHeader/wiki/%E9%85%8D%E7%BD%AE
+ * @FilePath: \webpack-demo\plugins\RmFilePlugin.js
+ */
 const path = require('path')
 const fs = require('fs')
 
@@ -18,17 +26,22 @@ class RmFilePlugin {
         compiler.hooks.beforeRun.tap('RmFilePlugin', (compiler) => {
             // 获取打包输出路径，（默认dist目录）
             const outputPath = webpackOptions.output.path || path.resolve(context, 'dist')
-            const fileList = fs.readdirSync(outputPath, {
-                withFileTypes: true
-            })
-            fileList.forEach(item => {
-                // 只删除文件，不对文件夹做递归删除，简化逻辑
-                if(item.isFile()) {
-                    const delPath = path.resolve(outputPath, item.name)
-                    console.log('删除文件===》',delPath)
-                    fs.unlinkSync(delPath)
+            fs.stat(outputPath, (err, stats) => {
+                if(!err && stats.isDirectory()) {
+                    const fileList = fs.readdirSync(outputPath, {
+                        withFileTypes: true
+                    })
+                    fileList.forEach(item => {
+                        // 只删除文件，不对文件夹做递归删除，简化逻辑
+                        if(item.isFile()) {
+                            const delPath = path.resolve(outputPath, item.name)
+                            console.log('删除文件===》',delPath)
+                            fs.unlinkSync(delPath)
+                        }
+                    })
                 }
             })
+            
         })
     }
 }
